Change references from chat to right-drawer
