+++ angular-app/src/app/app.component.html
@@ -6,8 +6,8 @@
     <mat-sidenav-content>
       <router-outlet></router-outlet>
     </mat-sidenav-content>
-    <mat-sidenav #chat class="sidechat-container" mode="side" position="end">
-      <app-chat-drawer></app-chat-drawer>
+    <mat-sidenav #rightDrawer class="right-drawer-container" mode="side" position="end">
+      <app-project-detail></app-project-detail>
     </mat-sidenav>
   </mat-sidenav-container>
 </div>
