+++ angular-app/src/app/project-detail/project-detail.component.html
@@ -5,8 +5,8 @@
     <button mat-icon-button matTooltip="Project Settings" (click)="openDialog(project)">
       <mat-icon>settings</mat-icon>
     </button>
-    <button mat-icon-button (click)="toggleChat()" aria-label="toggle chat" matTooltip="Toggle Chat">
-      <mat-icon>chat</mat-icon>
+    <button mat-icon-button (click)="toggleRightDrawer()">
+      <mat-icon>info</mat-icon>
     </button>
   </mat-toolbar>
   <mat-selection-list [multiple]="false">
