+++ angular-app/src/app/projects/projects.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { Project } from '../project';
-import { ToggleChatService } from '../toggle-chat.service';
 import { ProjectService } from '../project.service';
 
 @Component({
@@ -15,16 +14,10 @@ export class ProjectsComponent implements OnInit {
   selectedProject: Project;
 
   constructor(
-    private toggleChatService: ToggleChatService,
     private projectService: ProjectService
   ) { }
 
   ngOnInit(): void {
     this.projects$ = this.projectService.getProjects();
-  }
-
-  toggleChat(): void {
-    console.log("toggling chat");
-    this.toggleChatService.toggle();
   }
 }
