--- angular-app/src/app/toggle-chat.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ToggleChatService } from './toggle-chat.service';
-
-describe('ToggleChatService', () => {
-  let service: ToggleChatService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ToggleChatService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
