+++ angular-app/src/styles.css
@@ -30,8 +30,7 @@ body {
   height: 100vh;
 }
 
-.sidechat-container {
-  width: 400px;
+.right-drawer-container {
   height: 100vh;
 }
